feat:代码规范,多余内容去除
