+++ src/login-factory.ts
@@ -1,9 +1,8 @@
-
-import { RpcBase } from "lite-ts-ajax";
-import { AccountLogin } from "./ account-login";
-import { appLogin } from "./app-login";
-import { ILogin } from "./i-login";
-import { BuildLoginOption, LoginFactoryBase } from "./login-factory-base";
+import { RpcBase } from 'lite-ts-ajax';
+import { AccountLogin } from './account-login';
+import { appLogin } from './app-login';
+import { BuildLoginOption, LoginFactoryBase } from './login-factory-base';
+import { ILogin } from './i-login';
 
 export class LoginFactory extends LoginFactoryBase {
     public constructor(
@@ -13,8 +12,7 @@ export class LoginFactory extends LoginFactoryBase {
     }
 
     public build(opt: BuildLoginOption): ILogin {
-        const ctor = opt.account && AccountLogin
-            || opt.app && appLogin
+        const ctor = opt.account ? AccountLogin : appLogin;
         return new ctor(opt, this.m_Rpc);
     }
 }
