--- src/login-index.ts
@@ -1,1 +1,0 @@
-export * from "./i-login";
